+++ Core/NPulseLTZoneCommon.cpp
@@ -65,7 +65,7 @@ bool NLTZone::ADefault(void)
 // —брос процесса счета.
 bool NLTZone::AReset(void)
 {
- CachedNumAConnectors=static_cast<double>(GetNumAConnectors(0));
+ CachedNumAConnectors=static_cast<int>(GetNumAConnectors(0));
  return true;
 }
 // --------------------------
