+++ Core/NPulseMembrane.cpp
@@ -251,7 +251,7 @@ bool NPulseMembrane::ACalculate(void)
  Feedback=0;
 
  for(int i=0;i<NumInputs;i++)
-  for(size_t j=0;j<GetInputDataSize(i)[1];j++)
+  for(int j=0;j<GetInputDataSize(i)[1];j++)
    Feedback.v+=GetInputData(i)->Double[j];
 
  Feedback.v*=FeedbackGain.v;
